refactor(test): move tests for utils to unit folder
